refactor(layout): nest C sources under _c/ subsystem tree
Move the flat src/turbohtml/*.{c,h} into a subsystem-first _c/ tree
(core, tokenizer, dom, serialize, query/{css,xpath}, encoding, features,
data) and rewrite local includes to be subsystem-qualified. Pure
relocation: no code changes, all 30823 tests pass. Updates meson source
paths + include_directories, the clang-tidy header filter, the
clang-format generated-table excludes, and codegen output-path docs.
