+++ meson.build
@@ -17,24 +17,24 @@ py = import('python').find_installation(pure: false)
 py.extension_module(
     '_html',
     [
-        'src/turbohtml/_htmlmodule.c',
-        'src/turbohtml/annotation.c',
-        'src/turbohtml/charref.c',
-        'src/turbohtml/escape.c',
-        'src/turbohtml/linkify.c',
-        'src/turbohtml/links.c',
-        'src/turbohtml/markup.c',
-        'src/turbohtml/sanitize.c',
-        'src/turbohtml/token_type.c',
-        'src/turbohtml/tokenizer_sm.c',
-        'src/turbohtml/tokenizer_type.c',
-        'src/turbohtml/treebuilder.c',
-        'src/turbohtml/treebuilder_py.c',
-        'src/turbohtml/tree_type.c',
-        'src/turbohtml/unescape.c',
-        'src/turbohtml/xpath.c',
+        'src/turbohtml/_c/core/module.c',
+        'src/turbohtml/_c/tokenizer/charref.c',
+        'src/turbohtml/_c/tokenizer/statemachine.c',
+        'src/turbohtml/_c/tokenizer/token.c',
+        'src/turbohtml/_c/tokenizer/tokenizer.c',
+        'src/turbohtml/_c/dom/tree.c',
+        'src/turbohtml/_c/dom/binding.c',
+        'src/turbohtml/_c/dom/nodes.c',
+        'src/turbohtml/_c/serialize/escape.c',
+        'src/turbohtml/_c/serialize/unescape.c',
+        'src/turbohtml/_c/serialize/markup.c',
+        'src/turbohtml/_c/query/xpath/xpath.c',
+        'src/turbohtml/_c/features/sanitize.c',
+        'src/turbohtml/_c/features/linkify.c',
+        'src/turbohtml/_c/features/links.c',
+        'src/turbohtml/_c/features/annotation.c',
     ],
-    include_directories: include_directories('src/turbohtml'),
+    include_directories: include_directories('src/turbohtml/_c'),
     install: true,
     subdir: 'turbohtml',
 )
