+++ src/turbohtml/_c/data/entity_names.h
