+++ tools/generate_tlds.py
@@ -1,5 +1,5 @@
 """
-Generate src/turbohtml/tld_table.h from IANA's list of top-level domains.
+Generate src/turbohtml/_c/data/tld_table.h from IANA's list of top-level domains.
 
 linkify recognizes a bare domain like ``example.com`` as a link only when its last label is a real TLD, the same
 rule bleach used. The canonical list is IANA's, so this downloads it, lowercases and sorts every entry (including the
@@ -8,7 +8,7 @@ generated header shaped like ``tag_atom.h``: a sorted ``{name, len}`` array plus
 label is matched by bucketing on its first byte and a case-insensitive ``memcmp``. The IANA ``Version`` line is
 recorded in the header so a regeneration shows an auditable diff.
 
-Usage:  python tools/generate_tlds.py src/turbohtml/tld_table.h
+Usage:  python tools/generate_tlds.py src/turbohtml/_c/data/tld_table.h
 """
 
 from __future__ import annotations
